Construct ghc-pkg name from ghc filename
Replace 'ghc' with 'ghc-pkg' in the filename of the ghc executable. This
is better than assuming the executable starts with a 'ghc-' prefix, as
some distros provide a "triple" for the executable, e.g. 'x86_64-redhat-linux-ghc'.

Replacing 'ghc' with 'ghc-pkg' works for these distros as well and,
additionally, we don't have to manually take care of the '.exe' suffix.